# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication/Authorization Service",
        "title": "Sentinel",
        "contact": {},
        "license": {
            "name": "AGPL-3.0 (With additional terms, see NOTICE file)",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/.well-known/jwks.json": {
            "get": {
                "description": "RFC 7517 (https://datatracker.ietf.org/doc/html/rfc7517)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get JSON Web Keys (JWKs)",
                "operationId": "get-jwks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsebody.JWKs"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
        "/auth": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verify that user is logged-in",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verifies user authentication",
                "operationId": "verify",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token in Token Bearer format",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/userdto.Payload"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "490": {
                        "description": "User data desynchronization",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Token-Refresh-Required": {
                                "type": "string",
                                "description": "Set to 'true' when token refresh is required"
                            }
                        }
                    },
                    "491": {
                        "description": "Session revoked",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Session-Revoked": {
                                "type": "string",
                                "description": "Set to 'true' if current user session was revoked"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CSRF_Header": []
                    },
                    {
                        "CSRF_Cookie": []
                    }
                ],
                "description": "Create new access and refresh tokens and update current session info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refreshes auth tokens",
                "operationId": "refresh",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Refresh Token (sent as HTTP-Only cookie in actual requests)",
                        "name": "X-Refresh-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "491": {
                        "description": "Session revoked",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Session-Revoked": {
                                "type": "string",
                                "description": "Set to 'true' if current user session was revoked"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CSRF_Header": []
                    },
                    {
                        "CSRF_Cookie": []
                    }
                ],
                "description": "Login endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login into the service",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "User credentials and audience",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestbody.Auth"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "490": {
                        "description": "User data desynchronization",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Token-Refresh-Required": {
                                "type": "string",
                                "description": "Set to 'true' when token refresh is required"
                            }
                        }
                    },
                    "491": {
                        "description": "Session revoked",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Session-Revoked": {
                                "type": "string",
                                "description": "Set to 'true' if current user session was revoked"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRF_Header": []
                    },
                    {
                        "CSRF_Cookie": []
                    }
                ],
                "description": "Logout endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Revoke user session",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "490": {
                        "description": "User data desynchronization",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Token-Refresh-Required": {
                                "type": "string",
                                "description": "Set to 'true' when token refresh is required"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
        "/auth/csrf-token": {
            "get": {
                "description": "Generates CSRF token and sets correspond \"_csrf\" cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get CSRF token",
                "operationId": "get-csrf-token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responsebody.CSRF"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
...
        "/auth/sessions/{uid}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CSRF_Header": []
                    },
                    {
                        "CSRF_Cookie": []
                    }
                ],
                "description": "Revoke all existing non-revoked sessions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Revokes all user sessions",
                "operationId": "revoke-all-user-sessions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    },
                    "490": {
                        "description": "User data desynchronization",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Token-Refresh-Required": {
                                "type": "string",
                                "description": "Set to 'true' when token refresh is required"
                            }
                        }
                    },
                    "491": {
                        "description": "Session revoked",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        },
                        "headers": {
                            "X-Session-Revoked": {
                                "type": "string",
                                "description": "Set to 'true' if current user session was revoked"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responsebody.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requestbody.Auth": {
            "type": "object",
            "properties": {
                "audience": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "login": {
                    "type": "string",
                    "example": "admin@mail.com"
                },
                "password": {
                    "type": "string",
                    "example": "your-password"
                }
            }
        },
...
        "responsebody.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error"
                },
                "message": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
...
        "responsebody.JWKs": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responsebody.JSONWebKey"
                    }
                }
            }
        },
...
        "userdto.Payload": {
            "type": "object",
            "properties": {
                "audience": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "urn:api:auth",
                        "urn:api:billing",
                        "https://example.domain.com"
                    ]
                },
                "id": {
                    "type": "string",
                    "example": "d529a8d2-1eb4-4bce-82aa-e62095dbc653"
                },
                "login": {
                    "type": "string",
                    "example": "admin@mail.com"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user",
                        "moderator"
                    ]
                },
                "session-id": {
                    "type": "string",
                    "example": "35b92582-7694-4958-9751-1fef710cb94d"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token format: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CSRF_Cookie": {
            "description": "CSRF token in cookie. Required for state-changing requests, must match with token in X-CSRF-Token header",
            "type": "apiKey",
            "name": "_csrf",
            "in": "cookie"
        },
        "CSRF_Header": {
            "description": "CSRF token in request header. Required for state-changing requests, must match with token in cookie",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}